requirements.txt
flask
gunicorn
transformers
diffusers
accelerate
safetensors
torch
requests
pillow